backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic-settings==2.6.1
pydantic[email]==2.10.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
httpx==0.27.2
slowapi==0.1.9

# AI providers (install only what you use)
openai==1.57.0
anthropic==0.40.0
google-generativeai==0.8.3

# Vector DB + embeddings
chromadb==0.5.23
sentence-transformers==3.3.1

# Optional: Redis cache
redis[asyncio]==5.2.1
